Fix build configuration by casting `versionCode` to an integer in `app/build.gradle.kts`.

+++ app/build.gradle.kts
@@ -21,7 +21,7 @@ android {
         applicationId = "pl.oki.frostalert"
         minSdk = 26
         targetSdk = 36
-        versionCode = 2.1
+        versionCode = 2.1.toInt()
         versionName = "1.0"
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
